src/backend/ai-gateway/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
numpy==1.26.4
Pillow==10.3.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
structlog==24.1.0
